Added two fixed `Intersection2' methods from lib/coll.gi.
The fix is to replace IsFinite( C1 ) by
( HasIsFinite( C1 ) or CanComputeSize( C1 ) ) and IsFinite( C1 ),
respectively IsFinite( C2 ) by
( HasIsFinite( C2 ) or CanComputeSize( C2 ) ) and IsFinite( C2 ).

I noticed this in a case where the argument in question was an orbit object
representing an orbit which is probably infinite. The method tried to list
the points in the orbit, and ran out of memory.

Should the fix be applied to the respective methods in lib/coll.gi as well?
    SK