Change IsZ_pi from a property to a category
This seems correct, as there were no real methods to compute it. Also,
it is very similar to IsIntegers, which is also a category.

This makes a so-called "hidden" implication created by DeclareProperty
explicit, thereby resolving a bunch of warnings that show up if one
starts the upcoming GAP 4.11 with the `-N` command line option, and then
loads this package.

For some information on the background of this, see also
<https://github.com/gap-system/gap/issues/1649> and
<https://github.com/gap-system/gap/issues/2336>